requirements.txt
Flask==3.1.1
flask-cors==6.0.1
beautifulsoup4==4.13.4
requests==2.32.4
numpy==2.3.1
pandas==2.3.1
openai==1.54.3
google-generativeai==0.8.3
python-dotenv==1.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
click==8.2.1
colorama==0.4.6
itsdangerous==2.2.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
pytz==2025.2
python-dateutil==2.9.0.post0